// web-application/src/types-openapi/api/models/ListServiceRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListServiceRequest
 */
export interface ListServiceRequest {
    /**
     * 
     * @type {string}
     * @memberof ListServiceRequest
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListServiceRequest
     */
    status?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListServiceRequest
     */
    category?: string | null;
}

/**
 * Check if a given object implements the ListServiceRequest interface.
 */
export function instanceOfListServiceRequest(value: object): value is ListServiceRequest {
    return true;
}

export function ListServiceRequestFromJSON(json: any): ListServiceRequest {
    return ListServiceRequestFromJSONTyped(json, false);
}

export function ListServiceRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListServiceRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'status': json['status'] == null ? undefined : json['status'],
        'category': json['category'] == null ? undefined : json['category'],
    };
}

export function ListServiceRequestToJSON(json: any): ListServiceRequest {
    return ListServiceRequestToJSONTyped(json, false);
}

export function ListServiceRequestToJSONTyped(value?: ListServiceRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'status': value['status'],
        'category': value['category'],
    };
}